+++ src/utils.py
@@ -14,10 +14,7 @@ def resource_path(relative_path):
     if os.path.isabs(relative_path):
         return relative_path
     if getattr(sys, "frozen", False):
-        if sys.platform == "darwin":
-            base_path = os.path.join(os.path.dirname(sys.executable), "..", "Resources")
-        else:
-            base_path = sys._MEIPASS
+        base_path = sys._MEIPASS
     else:
         base_path = os.path.dirname(os.path.dirname(__file__))
     return os.path.normpath(os.path.join(base_path, relative_path))